Offsets int8_t type does not allow scrolling large matrixs
int8_t offset limits scrolling to 127 whereas it is supposed to be able to be incremented or decremented several times the pixel size of the initial buffer

+++ src/utility/LED_Display.cpp
@@ -90,7 +90,7 @@ void LED_Display::animation(uint8_t *buffptr, uint8_t amspeed, uint8_t ammode, i
     xSemaphoreGive(_xSemaphore);
 }
 
-void LED_Display::displaybuff(uint8_t *buffptr, int8_t offsetx, int8_t offsety)
+void LED_Display::displaybuff(uint8_t *buffptr, int32_t offsetx, int32_t offsety)
 {
     uint16_t xsize = 0, ysize = 0;
     xsize = buffptr[0];
